Proyect hierarchy refactor
