+++ app/src/test/java/com/baldev/eventify/domain/services/LoginServiceTest.java
@@ -1,4 +1,4 @@
-package com.baldev.eventify;
+package com.baldev.eventify.domain.services;
 
 import com.baldev.eventify.domain.actions.LoginCallback;
 import com.baldev.eventify.domain.entities.User;
